+++ src/main/java/com/vectorsearch/faiss/swig/IDSelector.java
@@ -35,7 +35,7 @@ public class IDSelector {
     }
   }
 
-  public boolean is_member(int id) {
+  public boolean is_member(long id) {
     return swigfaissJNI.IDSelector_is_member(swigCPtr, this, id);
   }
